# backend/openapi_doc/specs/openapi-v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Routine Cloud API v1",
    "description": "Routine Cloud API SDK",
    "version": "v1"
  },
  "paths": {
...
    "/runtime/active": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Active Runtime",
        "operationId": "runtime_active",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeActiveRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/sync": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Sync Runtime",
        "operationId": "runtime_sync",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeSyncRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/start": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Start Runtime",
        "operationId": "runtime_start",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCommandAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/pause": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Pause Runtime",
        "operationId": "runtime_pause",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCommandAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/resume": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Resume Runtime",
        "operationId": "runtime_resume",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCommandAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/skip": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Skip Runtime",
        "operationId": "runtime_skip",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCommandAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/stop": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Stop Runtime",
        "operationId": "runtime_stop",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCommandAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/complete": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Complete Runtime",
        "operationId": "runtime_complete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeCommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeCommandAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/runtime/events": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Stream Runtime Events",
        "operationId": "runtime_events",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-ID-Token",
            "in": "header",
            "required": false,
            "schema": {
              "title": "X-Id-Token",
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuntimeActiveRead": {
        "properties": {
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeStateRead"
          },
          "routine": {
            "$ref": "#/components/schemas/RuntimeRoutineRead",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "server_time",
          "runtime"
        ],
        "title": "RuntimeActiveRead"
      },
...
      "RuntimeCommandAccepted": {
        "properties": {
          "command_id": {
            "type": "string",
            "title": "Command Id"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "default": true
          },
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "sync": {
            "$ref": "#/components/schemas/RuntimeSyncRead"
          },
          "active": {
            "$ref": "#/components/schemas/RuntimeActiveRead",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "server_time",
          "sync"
        ],
        "title": "RuntimeCommandAccepted"
      },
      "RuntimeCommandRequest": {
        "properties": {
          "command_id": {
            "type": "string",
            "title": "Command Id"
          },
          "type": {
            "$ref": "#/components/schemas/RuntimeCommandType"
          },
          "routine_id": {
            "title": "Routine Id",
            "type": "integer",
            "nullable": true
          },
          "requested_at": {
            "title": "Requested At",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "source_device_id": {
            "title": "Source Device Id",
            "type": "string",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "type"
        ],
        "title": "RuntimeCommandRequest"
      },
      "RuntimeCommandType": {
        "type": "string",
        "enum": [
          "routine.start",
          "routine.pause",
          "routine.resume",
          "routine.skip",
          "routine.stop",
          "routine.complete"
        ],
        "title": "RuntimeCommandType"
      },
...
      "RuntimeRoutineRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "title": "Description",
            "type": "string",
            "nullable": true
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskInRoutineRead"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "tasks"
        ],
        "title": "RuntimeRoutineRead"
      },
      "RuntimeStateRead": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/RuntimeStatus"
          },
          "routine_id": {
            "title": "Routine Id",
            "type": "integer",
            "nullable": true
          },
          "participant_user_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Participant User Ids"
          },
          "current_task_id": {
            "title": "Current Task Id",
            "type": "integer",
            "nullable": true
          },
          "current_task_position": {
            "title": "Current Task Position",
            "type": "integer",
            "nullable": true
          },
          "task_started_at": {
            "title": "Task Started At",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "routine_started_at": {
            "title": "Routine Started At",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "paused_at": {
            "title": "Paused At",
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pause_duration": {
            "type": "integer",
            "title": "Pause Duration",
            "default": 0
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "RuntimeStateRead"
      },
      "RuntimeStatus": {
        "type": "string",
        "enum": [
          "idle",
          "running",
          "paused",
          "finished"
        ],
        "title": "RuntimeStatus"
      },
      "RuntimeSyncRead": {
        "properties": {
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time"
          },
          "runtime": {
            "$ref": "#/components/schemas/RuntimeStateRead"
          }
        },
        "type": "object",
        "required": [
          "server_time",
          "runtime"
        ],
        "title": "RuntimeSyncRead"
      },
...
      "TaskInRoutineRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "icon_name": {
            "type": "string",
            "title": "Icon Name"
          },
          "sound": {
            "type": "string",
            "title": "Sound"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "icon_name",
          "sound",
          "duration",
          "position"
        ],
        "title": "TaskInRoutineRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}